LAB2/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
pydantic==2.12.5
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11
python-dotenv==1.2.2
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
exceptiongroup==1.3.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
pip==26.0.1
pydantic_core==2.41.5
setuptools==57.4.0
starlette==0.52.1
tomli==2.4.0
typing_extensions==4.15.0
typing-inspection==0.4.2